refactor(manager): move node controls into its own module
